added custom hook for image url checking and displaying images in stack view now
